Refactor on tests module
